must avoid overflow for large particle counts. Use long instead of int.
(cherry picked from commit 0cd5314ece9a7023c08f9d93c595ba1d190e23ba)

+++ IO/H5part/vtkH5PartReader.cxx
@@ -606,8 +606,8 @@ int vtkH5PartReader::RequestData(vtkInformation* vtkNotUsed(request),
   {
     if (numPieces > 1)
     {
-      int div = Nt / numPieces;
-      int rem = Nt % numPieces;
+      long div = Nt / numPieces;
+      long rem = Nt % numPieces;
 
       vtkIdType myNt = piece < rem ? div + 1 : div;
       vtkIdType myOffset = piece < rem ? (div + 1) * piece : (div + 1) * rem + div * (piece - rem);
